--- src/main/kotlin/method/Management.kt
@@ -1,9 +1,0 @@
-package method
-
-interface Management {
-    //TODO: 1.Add
-    fun <T>addObject(data:T)
-
-    //TODO: 2.Read
-    fun <T>readObject(data:T)
-}
