+++ src/main/kotlin/method/util/GlobalMethod.kt
@@ -1,4 +1,4 @@
-package method
+package method.util
 
 import util.*
 import data.ObjectManagement
